backend/requirements.txt
pytesseract
Pillow
fastapi
uvicorn
python-multipart